+++ pkg/workers/mails/mail.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/cozy/cozy-stack/pkg/config"
 	"github.com/cozy/cozy-stack/pkg/instance"
+	"github.com/cozy/cozy-stack/pkg/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/jobs"
 	"github.com/cozy/cozy-stack/pkg/utils"
 	"github.com/cozy/gomail"
@@ -82,7 +83,7 @@ func SendMail(ctx *jobs.WorkerContext) error {
 		return err
 	}
 	domain := ctx.Domain()
-	i, err := instance.Get(domain)
+	i, err := lifecycle.GetInstance(domain)
 	if err != nil {
 		return err
 	}
