+++ web/dev.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/cozy/cozy-stack/pkg/config"
-	"github.com/cozy/cozy-stack/pkg/instance"
+	"github.com/cozy/cozy-stack/pkg/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/workers/mails"
 	"github.com/cozy/cozy-stack/web/middlewares"
 	"github.com/cozy/cozy-stack/web/statik"
@@ -78,7 +78,7 @@ func devData(c echo.Context) echo.Map {
 	if _, ok := data["ContextName"]; !ok {
 		data["ContextName"] = config.DefaultInstanceContext
 	}
-	if i, err := instance.Get(c.Request().Host); err == nil {
+	if i, err := lifecycle.GetInstance(c.Request().Host); err == nil {
 		data["CozyUI"] = middlewares.CozyUI(i)
 		data["ThemeCSS"] = middlewares.ThemeCSS(i)
 		data["Favicon"] = middlewares.Favicon(i)
